# openapi.yaml
openapi: 3.1.0
info:
  title: 27001certified API
  version: 1.0.0
  description: |
    Backend API for 27001certified.app — ISO 27001:2022 AI-Certification Portal.
    All protected endpoints require a JWT Bearer token unless marked as public.
  contact:
    name: 27001certified Platform
    url: https://27001certified.app
servers:
  - url: https://api.27001certified.app
    description: Production
  - url: https://api-staging.27001certified.app
    description: Staging
  - url: http://localhost:5000
    description: Local development

# ────────────────────────────────────────────────────────────
#  SECURITY SCHEMES
# ────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── SHARED SCHEMAS ─────────────────────────────────────────
  schemas:
...
    User:
      type: object
      properties:
        id:            { type: string, format: uuid }
        full_name:     { type: string }
        email:         { type: string, format: email }
        company:       { type: string, nullable: true }
        job_role:      { type: string, nullable: true }
        country:       { type: string, nullable: true }
        email_verified:{ type: boolean }
        role:          { type: string, enum: [user, admin] }
        created_at:    { type: string, format: date-time }
...
    Error:
      type: object
      required: [error]
      properties:
        error:   { type: string }
        message: { type: string }
        code:    { type: string }

    # ── Pagination ─────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Resource already exists or constraint violation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }

# ────────────────────────────────────────────────────────────
#  GLOBAL SECURITY (override per-route for public endpoints)
# ────────────────────────────────────────────────────────────
security:
  - BearerAuth: []

# ────────────────────────────────────────────────────────────
#  PATHS
# ────────────────────────────────────────────────────────────
paths:
...
  /api/auth/register:
    post:
      summary: Register a new user and send OTP
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [full_name, email, password]
              properties:
                full_name: { type: string, minLength: 2, maxLength: 100 }
                email:     { type: string, format: email }
                password:  { type: string, minLength: 8 }
                company:   { type: string, maxLength: 100 }
                job_role:  { type: string, maxLength: 100 }
                country:   { type: string, maxLength: 100 }
      responses:
        '201':
          description: User created, OTP sent to email
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:          { type: string }
                  user_id:          { type: string, format: uuid }
                  otp_expires_in_s: { type: integer, example: 600 }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/auth/verify-otp:
    post:
      summary: Verify OTP and receive JWT
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, code]
              properties:
                user_id: { type: string, format: uuid }
                code:    { type: string, pattern: '^[0-9]{6}$' }
      responses:
        '200':
          description: OTP verified — returns JWT and refresh token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:         { type: string }
                  expires_in:    { type: string, example: "24h" }
                  user:          { $ref: '#/components/schemas/User' }
          headers:
            Set-Cookie:
              schema:
                type: string
                description: HttpOnly refresh_token cookie (7d)
        '400':
          description: Invalid or expired OTP
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/auth/resend-otp:
    post:
      summary: Resend OTP (rate-limited 3/hr/email + 3/hr/IP)
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: OTP resent
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/auth/login:
    post:
      summary: Login with email + password
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:      { type: string }
                  expires_in: { type: string }
                  user:       { $ref: '#/components/schemas/User' }
          headers:
            Set-Cookie:
              schema: { type: string }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/auth/refresh:
    post:
      summary: Get new JWT using refresh token (HttpOnly cookie)
      security: []
      tags: [Auth]
      responses:
        '200':
          description: New JWT issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:      { type: string }
                  expires_in: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/forgot-password:
    post:
      summary: Send password reset link
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: Reset link sent (always returns 200 to prevent email enumeration)
  /api/auth/reset-password:
    post:
      summary: Reset password using signed token
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:    { type: string }
                password: { type: string, minLength: 8 }
      responses:
        '200':
          description: Password reset successfully
        '400':
          description: Invalid or expired reset token
  /api/auth/logout:
    post:
      summary: Revoke refresh token
      tags: [Auth]
      responses:
        '200':
          description: Logged out, refresh token revoked
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ══════════════════════════════════════════
  #  USER
  # ══════════════════════════════════════════
...